Remove preprocessor from `RestClient` build configuration.
