+++ CMakeLists.txt
@@ -384,29 +384,38 @@ set(TILEDB_EXPORT_HEADER "${TILEDB_EXPORT_HEADER_DIR}/${TILEDB_EXPORT_HEADER_NAM
 set(TILEDB_EXPORT_HEADER_LOCALINSTALL_PATH "tiledb/api/c_api/${TILEDB_EXPORT_HEADER_NAME}")
 
 ##################################
-# VFS Object Storage Configuration
+# Library Configuration
 ##################################
-# Create an interface library for use in passing the
-# appropriate compiler definitions for the set of
-# configured object storage backends.
+#
+# `configuration_definitions` is an interface library that contains compiler
+# definitions corresponding to the configuration variables passed to CMake
 
-add_library(object_store_definitions INTERFACE)
+add_library(configuration_definitions INTERFACE)
 
 if (TILEDB_AZURE)
-  target_compile_definitions(object_store_definitions INTERFACE -DHAVE_AZURE)
+  target_compile_definitions(configuration_definitions INTERFACE -DHAVE_AZURE)
 endif()
 
 if (TILEDB_GCS)
-  target_compile_definitions(object_store_definitions INTERFACE -DHAVE_GCS)
+  target_compile_definitions(configuration_definitions INTERFACE -DHAVE_GCS)
 endif()
 
 if (TILEDB_HDFS)
-  target_compile_definitions(object_store_definitions INTERFACE -DHAVE_HDFS)
+  target_compile_definitions(configuration_definitions INTERFACE -DHAVE_HDFS)
 endif()
 
 if (TILEDB_S3)
-  target_compile_definitions(object_store_definitions INTERFACE -DHAVE_S3)
+  target_compile_definitions(configuration_definitions INTERFACE -DHAVE_S3)
 endif()
+
+#
+# All object libraries should depend upon the `configuration_definitions`
+# library, but they don't at this time.
+# - The vfs library does not have all the definitions for all of the
+#   various object stores yet. This is an issue with the maturity of
+#   this object library only; it can be changed locally.
+#
+add_library(object_library_substitute_configuration_definitions INTERFACE)
 
 ##################################
 # Local install
