+++ tiledb/sm/rest/rest_client.h
@@ -27,7 +27,143 @@
  *
  * @section DESCRIPTION
  *
- * This file defines a REST client class.
+ * This file defines a pair of REST client classes: one disabled from
+ * interacting with a REST server and one enabled to. Construction of
+ * `RestClient` instances is available only through a factory.  Only one of
+ * these classes is available to be constructed, depending on the build
+ * configuration specified.
+ *
+ * @section Cyclic build dependency
+ *
+ * `class RestClient` is the nexus of a serious class of cyclic dependencies in
+ * the code. These cycles are a hard blocker to complete the agenda of
+ * converting the library to a hierarchy (i.e. DAG) of object libraries.
+ *
+ * The origin of the cycles is the commingling of REST operations with direct
+ * operations in the code of fundamental classes such as `Array`, `Query`, and
+ * `Group`. The skeleton of the dependency looks as follows; `Array::open` is
+ * the example.
+ *
+ *   - `Array::Array` is created with a URI. That URI can be either of two
+ *     possibilities. The member variable `remote_` is initialized accordingly.
+ *       - Direct URI to an object store or file system. (`remote_ == false`)
+ *       - REST URI that begins with `tiledb:`. (`remote_ == true`)
+ *   - Calling `Array::open` implements two different ways of opening the file,
+ *     depending on whether it's remote or not.
+ *       - Direct (non-remote) opening does not use `RestClient` in any way.
+ *         This operation does not create any cycles.
+ *       - Remote opening always uses `RestClient`, calling
+ *         `get_array_schema_from_rest` and `post_array_from_rest`.
+ *   - As a particular example, `post_array_from_rest` calls
+ *     `array_deserialize`. (All `RestClient` operations use capnp
+ *     serialization functions analogously.)
+ *   - `array_deserialize` calls the `Array` constructor.
+ *
+ * Compacting and summarizing the call sequence:
+ *   - `Array` depends on `RestClient` because of `Array::open` uses
+ *     `RestClient::post_array_from_rest`.
+ *   - `RestClient` depends on `Array` because of
+ *     `RestClient::post_array_from_rest` uses `Array::Array`.
+ *
+ * The root cause of the cycle is that the `Array` constructor commingles both
+ * direct and remote URIs into the same argument. As long as this persists,
+ * there's a necessary cycle. This example is a specific example of a widespread
+ * and generic pattern in the use of `RestClient`.
+ *
+ * The minimum requirement is that there be two separate constructors and that
+ * they be defined in different translation units (so that they don't end up in
+ * the same object file). These two constructors might both be on an existing
+ * class itself, possibly, or they might be on different classes. Regardless of
+ * the approach, the elimination of these cycles is a long term project.
+ *
+ * @section Dependency on TILEDB_SERIALIZATION
+ *
+ * The REST client has two different behaviors depending on the build
+ * configuration for serialization.
+ *   - If serialization is off, the REST client is a stub, throwing for all
+ *     remote operations.
+ *   - If serialization is on, the REST client performs remote operations
+ *     with a server.
+ *
+ * The earliest way this polymorphism was accomplished was with two different
+ * versions of the class, both named `class RestClient`, selected with the
+ * preprocessor. This was a distinctly C-style way of programming with a number
+ * of disadvantages. (All these are corollaries of the general maxim that The
+ * Preprocessor is Evil).
+ *   - Only a single configuration could be compiled in a build tree at a time.
+ *   - It was impossible to test both kinds of REST client with any single
+ *     configuration.
+ *   - All object libraries with any dependence on `class RestClient` came in
+ *     two different versions, but with the same name.
+ *
+ * In order to improve the situation, we have the following design goals:
+ *   - No use of the preprocessor in C++ code involving the REST client.
+ *   - For all code that is not involved in implementing the REST client, there
+ *     must be a single object library that applies to both configurations.
+ *   - When possible, all code involving implementing the server-enabled REST
+ *     client should be segregated into separate compilation units. This cannot
+ *     be a hard requirement until after all the dependency cycles are broken.
+ *   - When it's necessary that there be different object libraries that depend
+ *     upon the configuration, they must be have distinct names. The build
+ *     configuration variable may only be used to define how these object
+ *     libraries are used, and not how they are defined.
+ *   - Only C++ language mechanisms may be used to select configurations within
+ *     C++ code. These include the following:
+ *     - Class inheritance and virtual functions.
+ *     - Template instantiations. In particular, certain classes might usefully
+ *       be compiled with a boolean template argument for serialization.
+ *     - Link specifications.
+ *
+ * @section Alleviating trouble with `RestClient`
+ *
+ * Summarizing, we have two major issues with `RestClient`: dependency cycles
+ * and multiple configurations. There's an overlap between approach to address
+ * these issues, but there's one dominant principle that applies: Minimize the
+ * dependency of other code on `class RestClient`. This translates into a tiered
+ * approach:
+ *   - Almost all code references the REST client through a base class and its
+ *     virtual functions. This class does not depend upon the build
+ *     configuration in any way. As a consequence, only a small amount of code
+ *     has any involvement with this configuration variable.
+ *   - Construction of REST client objects occurs only through a factory
+ *     function. The factory function itself is necessarily involved with the
+ *     the build configuration, but also isolates outside code from how the
+ *     factory works.
+ *     - The only class that constructs REST client objects is `class Context`.
+ *       The factory eliminates a potential build dependency for this class.
+ *   - The factory perforce must behave in two modes: server-disabled and
+ *     server-enabled. Using two different classes for these two modes would
+ *     incur duplication in object libraries. Instead, the factory initializes
+ *     itself according the link specification.
+ *     - The constructor for server-disabled clients is always present. If
+ *       it's the only one present, that's what the factory returns.
+ *     - The constructor for server-enabled clients is optional; it may be
+ *       linked in or not. If it's linked it, the factory returns server-enabled
+ *       clients.
+ *
+ * This technical design makes a tradeoff. It incurs overhead in a few ways, all
+ * of which are neglible in practice.
+ * - There are virtual function calls for all REST client operations. Given that
+ *   each such operation performs I/O, the call overhead in minuscule. Put
+ *   differently, tight inner loops do not call these functions.
+ * - The code for the server-disabled REST client is present in the binary even
+ *   when the factory is producing server-enabled clients. The excess code is a
+ *   couple dozen stubs, all of which throw the same exception. The object code
+ *   overhead is less than a kilobyte.
+ * - The factory and its initialization have a layer of internal indirection.
+ *   Given that the only time REST clients are constructed is one per `class
+ *   Context`, this overhead is even less than that for its virtual functions.
+ *
+ * @section Upshot for users
+ *
+ * The resulting system is very easy to use:
+ * - If you want an executable with server-disabled objects, just link the code
+ *   you ordinarily would. Ensure your link specification does not inadvertently
+ *   include `rest_client_remote`.
+ * - If you want an executable with server-enabled objects, link it with the
+ *   object library `rest_client_remote`. The presence of this object library
+ *   changes the initialization of the factory from the default.
+ *
  */
 
 #ifndef TILEDB_REST_CLIENT_H
@@ -46,536 +182,378 @@ using namespace tiledb::common;
 
 namespace tiledb::sm {
 
+/**
+ * Exception class with `RestClient` origin
+ */
+class RestClientException : public StatusException {
+ public:
+  explicit RestClientException(const std::string& message)
+      : StatusException("RestClient", message) {
+  }
+};
+
+/**
+ * Exception class for operations disabled in the base class
+ */
+class RestClientDisabledException : public RestClientException {
+ public:
+  explicit RestClientDisabledException()
+      : RestClientException(
+            "Cannot use rest client; serialization not enabled.") {
+  }
+};
+
+/*
+ * Forward declarations required for `RestClient` member function declarations
+ */
 class ArraySchema;
 class ArraySchemaEvolution;
 class Config;
 class FragmentInfo;
-class Query;
 class MemoryTracker;
+class Query;
 class QueryPlan;
 
-enum class SerializationType : uint8_t;
+// Forward for friend declaration within `class RestClient`
+class RestClientFactory;
 
+/**
+ * Base class for REST client is also the server-disabled class.
+ *
+ * There are two REST client classes:
+ * 1. This one fails all operations. This is the "server-disabled" or "local"
+ *    version.
+ * 2. The other one interacts with a server to perform operations. That is the
+ *    "server-enabled" or "remote" version.
+ * The "local" vs. "remote" distinction is about where the processing happens,
+ * not where the storage is located. Their full names might be "locally-
+ * processed" vs. "remotely-processed", but that's not what they're called.
+ *
+ * This class (the local one) is the base class of the other (remote) one. As a
+ * result, this class does not have pure virtual methods. All virtual member
+ * functions defined in the base are implemented. They are, however, implemented
+ * as stubs that do nothing but throw.
+ *
+ * One facility is present in the base class: the extra headers. This implements
+ * the C API function `tiledb_ctx_set_tag`. In production code, this data is
+ * only used to additional HTTP headers. There are additional uses, however,
+ * including test code and context tracking. This facility is in the base class
+ * so that it's available regardless of the build configuration.
+ */
 class RestClient {
+  /*
+   * The factory is the only way to get to the constructor of this class.
+   */
+  friend class RestClientFactory;
+
+ protected:
+  /**
+   * The set of extra headers that are attached to REST requests. This is the
+   * data structure where the arguments to `tiledb_ctx_set_tag` end up.
+   */
+  std::unordered_map<std::string, std::string> extra_headers_;
+
+  /**
+   * Name of the rest server used for remote operations.
+   */
+  std::string rest_server_;
+
  public:
-  /** Constructor. */
-  RestClient();
-
-  /** Initialize the REST client with the given config. */
-  Status init(
-      stats::Stats* parent_stats,
-      const Config* config,
-      ThreadPool* compute_tp,
-      const std::shared_ptr<Logger>& logger,
-      ContextResources& resources);
-
-  /** Sets a header that will be attached to all requests. */
-  Status set_header(const std::string& name, const std::string& value);
+  RestClient(const Config& config);
 
   /**
-   * Check if use_refactored_array_open_and_query_submit is set in
-   * input config so that rest_client chooses the right URI
+   * Add a header that will be attached to all requests.
    *
-   * @param config Config to check
-   *
-   * */
-  static bool use_refactored_query(const Config& config);
-
-  /**
-   * Check if an array exists by making a REST call. To start with this fetches
-   * the schema but ignores the body returned if non-error
-   *
-   * @param uri to check
-   * @return tuple of Status and if array exists
+   * This member function supports `tiledb_ctx_set_tag`. Although the designated
+   * use of this function is for remote REST operations, it needs to be
+   * available in both local and remote configurations for testing.
    */
-  tuple<Status, std::optional<bool>> check_array_exists_from_rest(
-      const URI& uri);
-
-  /**
-   * Check if an group exists by making a REST call. To start with this fetches
-   * the schema but ignores the body returned if non-error
-   *
-   * @param uri to check
-   * @return tuple of Status and if group exists
-   */
-  tuple<Status, std::optional<bool>> check_group_exists_from_rest(
-      const URI& uri);
-
-  /**
-   * Get a data encoded array schema from rest server.
-   *
-   * @param uri of array being loaded
-   * @return Status and new ArraySchema shared pointer.
-   */
-  tuple<Status, optional<shared_ptr<ArraySchema>>> get_array_schema_from_rest(
-      const URI& uri);
-
-  /**
-   * Get an array schema from the rest server. This will eventually replace the
-   * get_array_schema_from_rest after TileDB-Cloud-REST merges support for the
-   * POST endpoint.
-   *
-   * @param uri The Array URI to load the schema from.
-   * @return shared_ptr<ArraySchema> The loaded array schema.
-   */
-  shared_ptr<ArraySchema> post_array_schema_from_rest(
-      const Config& config,
-      const URI& uri,
-      uint64_t timestamp_start,
-      uint64_t timestamp_end,
-      bool include_enumerations = false);
-
-  /**
-   * Post the array config and get an array from rest server
-   *
-   * @param uri of array being loaded
-   * @param resources the context resources
-   * @param array array to load into
-   */
-  void post_array_from_rest(
-      const URI& uri, ContextResources& resources, Array* array);
-
-  /**
-   * Post a data array schema to rest server
-   *
-   * @param uri of array being created
-   * @param array_schema array schema to load into
-   * @return Status Ok() on success Error() on failures
-   */
-  Status post_array_schema_to_rest(
-      const URI& uri, const ArraySchema& array_schema);
-
-  /**
-   * Deletes all written data from array at the given URI from the REST server.
-   *
-   * @param uri Array URI to delete
-   */
-  void delete_array_from_rest(const URI& uri);
-
-  /**
-   * Deletes the fragments written between the given timestamps from the array
-   * at the given URI from the REST server.
-   *
-   * @param uri Array URI to delete fragments from
-   * @param array Array to delete fragments from
-   * @param timestamp_start The start timestamp at which to delete fragments
-   * @param timestamp_end The end timestamp at which to delete fragments
-   *
-   * #TODO Implement API endpoint on TileDBCloud.
-   */
-  void post_delete_fragments_to_rest(
-      const URI& uri,
-      Array* array,
-      uint64_t timestamp_start,
-      uint64_t timestamp_end);
-
-  /**
-   * Deletes the fragments with the given URIs from the array at the given URI
-   * from the REST server.
-   *
-   * @param uri Array URI to delete fragments from
-   * @param array Array to delete fragments from
-   * @param fragment_uris The uris of the fragments to be deleted
-   *
-   * #TODO Implement API endpoint on TileDBCloud.
-   */
-  void post_delete_fragments_list_to_rest(
-      const URI& uri, Array* array, const std::vector<URI>& fragment_uris);
-
-  /**
-   * Deregisters an array at the given URI from the REST server.
-   *
-   * @param uri Array URI to deregister
-   * @return Status
-   */
-  Status deregister_array_from_rest(const URI& uri);
-
-  /**
-   * Get array's non_empty domain from rest server
-   *
-   * @param array Array model to fetch and set non empty domain on
-   * @param timestamp_start Inclusive starting timestamp at which to open array
-   * @param timestamp_end Inclusive ending timestamp at which to open array
-   * @return Status Ok() on success Error() on failures
-   */
-  Status get_array_non_empty_domain(
-      Array* array, uint64_t timestamp_start, uint64_t timestamp_end);
-
-  /**
-   * Get array's max buffer sizes from rest server.
-   *
-   * @param uri URI of array
-   * @param schema Array schema of array
-   * @param subarray Subrray to get max buffer sizes for
-   * @param buffer_sizes Will be populated with max buffer sizes
-   * @return Status
-   */
-  Status get_array_max_buffer_sizes(
-      const URI& uri,
-      const ArraySchema& schema,
-      const void* subarray,
-      std::unordered_map<std::string, std::pair<uint64_t, uint64_t>>*
-          buffer_sizes);
-
-  /**
-   * Gets the array's metadata from the REST server (and updates the in-memory
-   * Metadata of the array to match the returned values).
-   *
-   * @param uri Array URI
-   * @param timestamp_start Inclusive starting timestamp at which to open array
-   * @param timestamp_end Inclusive ending timestamp at which to open array
-   * @param array Array to fetch metadata for
-   * @return Status
-   */
-  Status get_array_metadata_from_rest(
-      const URI& uri,
-      uint64_t timestamp_start,
-      uint64_t timestamp_end,
-      Array* array);
-
-  /**
-   * Posts the array's metadata to the REST server.
-   *
-   * @param uri Array URI
-   * @param timestamp_start Inclusive starting timestamp at which to open array
-   * @param timestamp_end Inclusive ending timestamp at which to open array
-   * @param array Array to update/post metadata for.
-   * @return Status
-   */
-  Status post_array_metadata_to_rest(
-      const URI& uri,
-      uint64_t timestamp_start,
-      uint64_t timestamp_end,
-      Array* array);
-
-  /**
-   * Get the requested enumerations from the REST server via POST request.
-   *
-   * @param uri Array URI.
-   * @param timestamp_start Inclusive starting timestamp at which to open array.
-   * @param timestamp_end Inclusive ending timestamp at which to open array.
-   * @param array Array to fetch metadata for.
-   * @param enumeration_names The names of the enumerations to get.
-   */
-  std::vector<shared_ptr<const Enumeration>> post_enumerations_from_rest(
-      const URI& uri,
-      uint64_t timestamp_start,
-      uint64_t timestamp_end,
-      Array* array,
-      const std::vector<std::string>& enumeration_names,
-      shared_ptr<MemoryTracker> memory_tracker = nullptr);
-
-  /**
-   * Get the requested query plan from the REST server via POST request.
-   *
-   * @param uri Array URI.
-   * @param query Query to fetch query plan for.
-   * @param query_plan The requested query plan.
-   */
-  void post_query_plan_from_rest(
-      const URI& uri, Query& query, QueryPlan& query_plan);
-
-  /**
-   * Post a data query to rest server
-   *
-   * @param uri of array being queried
-   * @param query to send to server and store results in, this qill be modified
-   * @return Status Ok() on success Error() on failures
-   */
-  Status submit_query_to_rest(const URI& uri, Query* query);
-
-  /**
-   * Post a data query to rest server
-   *
-   * @param uri of array being queried
-   * @param query to send to server and store results in, this will be modified
-   * @return Status Ok() on success Error() on failures
-   */
-  Status finalize_query_to_rest(const URI& uri, Query* query);
-
-  /**
-   * Submit and finalize a query to rest server. Used in global order
-   * writes to submit the last tile-unaligned chunk and finalize the query.
-   *
-   * @param uri of array being queried
-   * @param query to send to server and store results in
-   * @return Status Ok() on success Error() on failures
-   */
-  Status submit_and_finalize_query_to_rest(const URI& uri, Query* query);
-
-  /**
-   * Get array's non_empty domain from rest server
-   *
-   * @param array Array model to fetch and set non empty domain on
-   * @return Status Ok() on success Error() on failures
-   */
-  Status get_query_est_result_sizes(const URI& uri, Query* query);
-
-  /**
-   * Post array schema evolution to rest server
-   *
-   * @param uri of array being queried
-   * @param array_schema_evolution to send to server
-   * @return Status Ok() on success Error() on failures
-   */
-  Status post_array_schema_evolution_to_rest(
-      const URI& uri, ArraySchemaEvolution* array_schema_evolution);
-
-  /**
-   * Get array's fragment info from rest server
-   *
-   * @param uri Array uri to query for
-   * @param fragment_info Fragment info object to store the incoming info
-   * @return Status Ok() on success Error() on failures
-   */
-  Status post_fragment_info_from_rest(
-      const URI& uri, FragmentInfo* fragment_info);
-
-  /**
-   * Gets the group's metadata from the REST server (and updates the in-memory
-   * Metadata of the group to match the returned values).
-   *
-   * @param uri Group URI
-   * @param group Group to fetch metadata for
-   * @return Status
-   */
-  Status post_group_metadata_from_rest(const URI& uri, Group* group);
-
-  /**
-   * Posts the group's metadata to the REST server.
-   *
-   * @param uri Group URI
-   * @param group Group to update/post metadata for.
-   * @return Status
-   */
-  Status put_group_metadata_to_rest(const URI& uri, Group* group);
-
-  /**
-   * Get group details from the REST server.
-   *
-   * @param uri Group UI
-   * @param group Group to deserialize into
-   * @return Status
-   */
-  Status post_group_from_rest(const URI& uri, Group* group);
-
-  /**
-   * Post group details to the REST server.
-   *
-   * @param uri Group UI
-   * @param group Group to serialize
-   * @return Status
-   */
-  Status patch_group_to_rest(const URI& uri, Group* group);
-
-  /**
-   * Deletes all written data from group at the given URI from the REST server.
-   *
-   * @param uri Group URI to delete
-   * @param recursive True if all data inside the group is to be deleted
-   */
-  void delete_group_from_rest(const URI& uri, bool recursive);
-
-  /**
-   * Post group create to the REST server.
-   *
-   * @param uri Group UI
-   * @param group Group to create
-   * @return Status
-   */
-  Status post_group_create_to_rest(const URI& uri, Group* group);
-
-  /**
-   * Post array consolidation request to the REST server.
-   *
-   * @param uri Array URI
-   * @param config config
-   * @return
-   */
-  Status post_consolidation_to_rest(const URI& uri, const Config& config);
-
-  /**
-   * Post array vacuum request to the REST server.
-   *
-   * @param uri Array URI
-   * @param config config
-   * @return
-   */
-  Status post_vacuum_to_rest(const URI& uri, const Config& config);
-
-  inline std::string rest_server() const {
-    return rest_server_;
-  }
-
-  /**
-   * Get consolidation plan from the REST server via POST request.
-   *
-   * @param uri Array URI.
-   * @param config Config of the array.
-   * @param fragment_size Maximum fragment size for constructing the plan.
-   * @return The requested consolidation plan
-   */
-  std::vector<std::vector<std::string>> post_consolidation_plan_from_rest(
-      const URI& uri, const Config& config, uint64_t fragment_size);
+  void add_header(const std::string& name, const std::string& value);
 
   /**
    * Constant accessor to `extra_headers_` member variable.
+   *
+   * This function supports testing the C API function `tiledb_ctx_set_tag`,
+   * which is write-only with respect to the C API.
    */
   const std::unordered_map<std::string, std::string>& extra_headers() const {
     return extra_headers_;
   }
 
- private:
-  /* ********************************* */
-  /*        PRIVATE ATTRIBUTES         */
-  /* ********************************* */
+  /**
+   * Accessor to configured rest server
+   *
+   * @section Maturity
+   *
+   * This function only make sense for remote operation. It's in the base class
+   * at present to avoid rewriting tests.
+   */
+  inline std::string rest_server() const {
+    return rest_server_;
+  }
 
-  /** The class stats. */
-  stats::Stats* stats_;
+  //-------------------------------------------------------
+  // Rest client operations
+  //-------------------------------------------------------
 
-  /** The TileDB config options (contains server and auth info). */
-  const Config* config_;
+  /// Operation disabled in base class.
+  inline virtual tuple<Status, std::optional<bool>>
+  check_array_exists_from_rest(const URI&) {
+    throw RestClientDisabledException();
+  }
 
-  /** The thread pool for compute-bound tasks. */
-  ThreadPool* compute_tp_;
+  /// Operation disabled in base class.
+  inline virtual tuple<Status, std::optional<bool>>
+  check_group_exists_from_rest(const URI&) {
+    throw RestClientDisabledException();
+  }
 
-  /** Rest server config param. */
-  std::string rest_server_;
+  /// Operation disabled in base class.
+  inline virtual tuple<Status, optional<shared_ptr<ArraySchema>>>
+  get_array_schema_from_rest(const URI&) {
+    throw RestClientDisabledException();
+  }
 
-  /** Serialization type. */
-  SerializationType serialization_type_;
+  /// Operation disabled in base class.
+  inline virtual shared_ptr<ArraySchema> post_array_schema_from_rest(
+      const Config&, const URI&, uint64_t, uint64_t, bool) {
+    throw RestClientDisabledException();
+  }
+
+  /// Operation disabled in base class.
+  inline virtual void post_array_from_rest(
+      const URI&, ContextResources&, Array*) {
+    throw RestClientDisabledException();
+  }
+
+  /// Operation disabled in base class.
+  inline virtual Status post_array_schema_to_rest(
+      const URI&, const ArraySchema&) {
+    throw RestClientDisabledException();
+  }
+
+  /// Operation disabled in base class.
+  inline virtual void delete_array_from_rest(const URI&) {
+    throw RestClientDisabledException();
+  }
+
+  /// Operation disabled in base class.
+  inline virtual void post_delete_fragments_to_rest(
+      const URI&, Array*, uint64_t, uint64_t) {
+    throw RestClientDisabledException();
+  }
+
+  /// Operation disabled in base class.
+  inline virtual void post_delete_fragments_list_to_rest(
+      const URI&, Array*, const std::vector<URI>&) {
+    throw RestClientDisabledException();
+  }
+
+  /// Operation disabled in base class.
+  inline virtual Status deregister_array_from_rest(const URI&) {
+    throw RestClientDisabledException();
+  }
+
+  /// Operation disabled in base class.
+  inline virtual Status get_array_non_empty_domain(Array*, uint64_t, uint64_t) {
+    throw RestClientDisabledException();
+  }
+
+  /// Operation disabled in base class.
+  inline virtual Status get_array_max_buffer_sizes(
+      const URI&,
+      const ArraySchema&,
+      const void*,
+      std::unordered_map<std::string, std::pair<uint64_t, uint64_t>>*) {
+    throw RestClientDisabledException();
+  }
+
+  /// Operation disabled in base class.
+  inline virtual Status get_array_metadata_from_rest(
+      const URI&, uint64_t, uint64_t, Array*) {
+    throw RestClientDisabledException();
+  }
+
+  /// Operation disabled in base class.
+  inline virtual Status post_array_metadata_to_rest(
+      const URI&, uint64_t, uint64_t, Array*) {
+    throw RestClientDisabledException();
+  }
+
+  /// Operation disabled in base class.
+  inline virtual std::vector<shared_ptr<const Enumeration>>
+  post_enumerations_from_rest(
+      const URI&,
+      uint64_t,
+      uint64_t,
+      Array*,
+      const std::vector<std::string>&,
+      shared_ptr<MemoryTracker> memory_tracker = nullptr) {
+    throw RestClientDisabledException();
+  }
+
+  /// Operation disabled in base class.
+  inline virtual void post_query_plan_from_rest(
+      const URI&, Query&, QueryPlan&) {
+    throw RestClientDisabledException();
+  }
+
+  /// Operation disabled in base class.
+  inline virtual Status submit_query_to_rest(const URI&, Query*) {
+    throw RestClientDisabledException();
+  }
+
+  /// Operation disabled in base class.
+  inline virtual Status finalize_query_to_rest(const URI&, Query*) {
+    throw RestClientDisabledException();
+  }
+
+  /// Operation disabled in base class.
+  inline virtual Status submit_and_finalize_query_to_rest(const URI&, Query*) {
+    throw RestClientDisabledException();
+  }
+
+  /// Operation disabled in base class.
+  inline virtual Status get_query_est_result_sizes(const URI&, Query*) {
+    throw RestClientDisabledException();
+  }
+
+  /// Operation disabled in base class.
+  inline virtual Status post_array_schema_evolution_to_rest(
+      const URI&, ArraySchemaEvolution*) {
+    throw RestClientDisabledException();
+  }
+
+  /// Operation disabled in base class.
+  inline virtual Status post_fragment_info_from_rest(
+      const URI&, FragmentInfo*) {
+    throw RestClientDisabledException();
+  }
+
+  /// Operation disabled in base class.
+  inline virtual Status post_group_metadata_from_rest(const URI&, Group*) {
+    throw RestClientDisabledException();
+  }
+
+  /// Operation disabled in base class.
+  inline virtual Status put_group_metadata_to_rest(const URI&, Group*) {
+    throw RestClientDisabledException();
+  }
+
+  /// Operation disabled in base class.
+  inline virtual Status post_group_from_rest(const URI&, Group*) {
+    throw RestClientDisabledException();
+  }
+
+  /// Operation disabled in base class.
+  inline virtual Status patch_group_to_rest(const URI&, Group*) {
+    throw RestClientDisabledException();
+  }
+
+  /// Operation disabled in base class.
+  inline virtual void delete_group_from_rest(const URI&, bool) {
+    throw RestClientDisabledException();
+  }
+
+  /// Operation disabled in base class.
+  inline virtual Status post_group_create_to_rest(const URI&, Group*) {
+    throw RestClientDisabledException();
+  }
+
+  /// Operation disabled in base class.
+  inline virtual Status post_consolidation_to_rest(const URI&, const Config&) {
+    throw RestClientDisabledException();
+  }
+
+  /// Operation disabled in base class.
+  inline virtual Status post_vacuum_to_rest(const URI&, const Config&) {
+    throw RestClientDisabledException();
+  }
+
+  /// Operation disabled in base class.
+  inline virtual std::vector<std::vector<std::string>>
+  post_consolidation_plan_from_rest(const URI&, const Config&, uint64_t) {
+    throw RestClientDisabledException();
+  }
+};
+
+// Forward declaration.
+class RestClientFactoryAssistant;
+
+/**
+ * Factory class for `RestClient`.
+ */
+class RestClientFactory {
+  /*
+   * Friend declaration allows override of the default factory behavior. The
+   * assistant has disciplined interaction with the internals of this class,
+   * unlike arbitrary friends might be.
+   */
+  friend class RestClientFactoryAssistant;
 
   /**
-   * If true (the default), automatically resubmit incomplete queries on the
-   * server-side. This guarantees that the user only receive a complete query
-   * result from the server.
-   *
-   * When this is turned on, it is currently an error if the user buffers on the
-   * client are too small to receive all data received from the server
-   * (regardless of how many times the query is resubmitted).
+   * Alias for the function type of the factory.
    */
-  bool resubmit_incomplete_;
-
-  /** Collection of extra headers that are attached to REST requests. */
-  std::unordered_map<std::string, std::string> extra_headers_;
-
-  /** Array URI to redirected server mapping. */
-  std::unordered_map<std::string, std::string> redirect_meta_;
-
-  /** Mutex for thread-safety. */
-  mutable std::mutex redirect_mtx_;
-
-  /** The class logger. */
-  shared_ptr<Logger> logger_;
-
-  /** UID of the logger instance */
-  inline static std::atomic<uint64_t> logger_id_ = 0;
-
-  /** The class MemoryTracker. */
-  shared_ptr<MemoryTracker> memory_tracker_;
-
-  /* ********************************* */
-  /*         PRIVATE METHODS           */
-  /* ********************************* */
+  using factory_type = std::shared_ptr<RestClient>(
+      stats::Stats& parent_stats,
+      const Config& config,
+      ThreadPool& compute_tp,
+      Logger& logger,
+      shared_ptr<MemoryTracker>&& tracker);
 
   /**
-   * POSTs a query submit request to the REST server and deserializes the
-   * response into the same query object.
    *
-   * For read queries, this also updates the given copy state with the number of
-   * bytes copied for each attribute, which allows for automatic resubmission of
-   * incomplete queries while concatenating to the user buffers.
-   *
-   * @param uri URI of array being queried
-   * @param query Query to send to server and store results in, this will be
-   *    modified.
-   * @param copy_state Map of copy state per attribute. As attribute data is
-   *    copied into user buffers on reads, the state of each attribute in this
-   *    map is updated accordingly.
-   * @return
    */
-  Status post_query_submit(
-      const URI& uri, Query* query, serialization::CopyState* copy_state);
+  static factory_type* factory_override_;
 
+ public:
   /**
-   * Callback to invoke as partial, buffered response data is received from
-   * posting a query.
+   * Factory for `RestClient` objects
    *
-   * This is not thread-safe. It expects the response data to be ordered. The
-   * response must contain serialized query objects, prefixed by an 8-byte
-   * unsigned integer that contains the byte-size of the serialized query object
-   * it is prefixing. The scratch space must be empty before the first
-   * invocation, and must not change until the last invocation has completed.
+   * @section Maturity
    *
-   * @param reset True if the callback must wipe the in-memory state
-   * @param contents the partial response data
-   * @param content_nbytes the size of the response data in 'contents'
-   * @param skip_retries Output argument that can be set to true to
-   *    prevent the curl layer from retrying this request.
-   * @scratch scratch space to use between invocations of this callback
-   * @query the query object used for deserializing the serialized query
-   *    objects in the response data.
-   * @param copy_state Map of copy state per attribute. As attribute data is
-   *    copied into user buffers on reads, the state of each attribute in this
-   *    map is updated accordingly.
-   * @return Number of acknowledged bytes
+   * The construction arguments are all the elements of `ContextResources`
+   * except one: the `RestClient` instance. The `RestClient` is current part
+   * of the context resources rather than part of the context itself. Once
+   * the member variable moves from `ContextResources` to `Context`, the
+   * factory can simply take a single `ContextResources` object.
+   *
+   * @param parent_stats The stats object from `ContextResources`
+   * @param config  The config from `ContextResources`
+   * @param compute_tp The compute thread pool from `ContextResources`
+   * @param logger The logger from `ContextResources`
+   * @param tracker The memory tracker from `ContextResources`.
+   * @return a newly constructed instance
    */
-  size_t query_post_call_back(
-      const bool reset,
-      void* constcontents,
-      const size_t content_nbytes,
-      bool* constskip_retries,
-      shared_ptr<Buffer> scratch,
-      Query* query,
-      serialization::CopyState* copy_state);
+  static std::shared_ptr<RestClient> make(
+      stats::Stats& parent_stats,
+      const Config& config,
+      ThreadPool& compute_tp,
+      Logger& logger,
+      shared_ptr<MemoryTracker>&& tracker);
+};
 
-  /**
-   * Returns a string representation of the given subarray. The format is:
-   *
-   *   "dim0min,dim0max,dim1min,dim1max,..."
-   *
-   * @param schema Array schema to use for domain information
-   * @param subarray Subarray to convert to string
-   * @param subarray_str Will be set to the CSV string
-   * @return Status
-   */
-  static Status subarray_to_str(
-      const ArraySchema& schema,
-      const void* subarray,
-      std::string* subarray_str);
+/*
+ * Forward declaration of the remote class.
+ *
+ * This declaration is necessary for a friend declaration in the factory. For
+ * the most part we want all references to the REST client to be through the
+ * base class and not refer to the remote class specifically.
+ */
+class RestClientRemoteFactory;
 
-  /**
-   * Sets the buffer sizes on the given query using the given state mapping (per
-   * attribute). Applicable only when deserializing read queries on the client.
-   *
-   * @param copy_state State map of attribute to buffer sizes.
-   * @param query Query to update buffers for
-   * @return Status
-   */
-  Status update_attribute_buffer_sizes(
-      const serialization::CopyState& copy_state, Query* query) const;
+/**
+ * Assistant to the factory provides a limited interface to the internals of
+ * the factory.
+ *
+ * This class has no public members. Its functions are only available to
+ * friends.
+ */
+class RestClientFactoryAssistant {
+  friend class RestClientRemoteFactory;
+  using factory_type = RestClientFactory::factory_type;
 
-  /**
-   * Helper function encapsulating the functionality of looking up for cached
-   * redirected rest server addresses to avoid the redirection overhead
-   *
-   * @param array_ns Array namespace
-   * @param array_uri Array URI
-   * @return Returns the redirection URI if exists and empty string otherwise
-   */
-  std::string redirect_uri(const std::string& cache_key);
-
-  /**
-   * Cap'n proto requires JSON messages to be null terminated c-style strings.
-   * This function checks if using JSON that returned buffers are null delimited
-   *
-   * @param buffer of server message
-   * @return Status
-   */
-  Status ensure_json_null_delimited_string(Buffer* buffer);
-
-  /** Load all custom headers from the given config. */
-  void load_headers(const Config& cfg);
+  static factory_type* override_factory(factory_type* f);
 };
 
 }  // namespace tiledb::sm
