+++ tiledb/sm/storage_manager/storage_manager.cc
@@ -105,8 +105,6 @@ Status StorageManagerCanonical::init() {
   auto& global_state = global_state::GlobalState::GetGlobalState();
   global_state.init(config_);
 
-  RETURN_NOT_OK(set_default_tags());
-
   global_state.register_storage_manager(this);
 
   return Status::Ok();
@@ -188,11 +186,6 @@ void StorageManagerCanonical::decrement_in_progress() {
   queries_in_progress_cv_.notify_all();
 }
 
-const std::unordered_map<std::string, std::string>&
-StorageManagerCanonical::tags() const {
-  return tags_;
-}
-
 void StorageManagerCanonical::increment_in_progress() {
   std::unique_lock<std::mutex> lck(queries_in_progress_mtx_);
   queries_in_progress_++;
@@ -212,17 +205,6 @@ Status StorageManagerCanonical::query_submit_async(Query* query) {
   return async_push_query(query);
 }
 
-Status StorageManagerCanonical::set_tag(
-    const std::string& key, const std::string& value) {
-  tags_[key] = value;
-
-  // Tags are added to REST requests as HTTP headers.
-  if (resources_.rest_client() != nullptr)
-    throw_if_not_ok(resources_.rest_client()->set_header(key, value));
-
-  return Status::Ok();
-}
-
 void StorageManagerCanonical::wait_for_zero_in_progress() {
   std::unique_lock<std::mutex> lck(queries_in_progress_mtx_);
   queries_in_progress_cv_.wait(
@@ -232,16 +214,5 @@ void StorageManagerCanonical::wait_for_zero_in_progress() {
 /* ****************************** */
 /*         PRIVATE METHODS        */
 /* ****************************** */
-
-Status StorageManagerCanonical::set_default_tags() {
-  const auto version = std::to_string(constants::library_version[0]) + "." +
-                       std::to_string(constants::library_version[1]) + "." +
-                       std::to_string(constants::library_version[2]);
-
-  RETURN_NOT_OK(set_tag("x-tiledb-version", version));
-  RETURN_NOT_OK(set_tag("x-tiledb-api-language", "c"));
-
-  return Status::Ok();
-}
 
 }  // namespace tiledb::sm
